(MODULES-10638) Temporarily remove breaking unit test
During some onceover executions of the Chocolatey rspec suite, the tests
for the chocolatey_install helper seem to run unguarded on Linux systems,
ignoring the guards in place.

While effort has not been made at this time to trace how this is happening
from within a onceover execution, execution of the spec tests outside of
onceover has shown the guards working.

For the time being, this commit temporarily removes the offending unit tests
from the module for the purposes of unblocking onceover users who are
experiencing this issue.

This branch will be updated with each release of the module until an actual
fix is diagnosed and put in place, at which time this branch will be deleted.